Explicitly delineate public vs. private methods
There is one method which is tested in the specifications that also depends on other methods for implementation. This method, Rectangle#generate_face_milling_toolpath, is public, and the other Rectangle instance methods that it depends on are private. The CNCHelper instance methods are unchanged since they do not depend on any other methods for implementation.